Added BankTransaction, BankTransactionText persistence classes and associated unit and integration tests.

+++ src/org/mariangolea/fintrack/bank/parser/persistence/transactions/BankTransactionTextRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BankTransactionTextRepository  extends JpaRepository<BankTransactionText, Integer>{
+public interface BankTransactionTextRepository  extends JpaRepository<BankTransactionText, Long>{
     
 }
